plan added all functions

+++ src/app/models/Plan.js
@@ -6,13 +6,12 @@ class Plan extends Model {
       {
         title: Sequelize.STRING,
         duration: Sequelize.INTEGER,
-        price: Sequelize.INTEGER,
+        price: Sequelize.FLOAT,
       },
       {
         sequelize,
       }
     );
-
     return this;
   }
 }
